+++ src/cli/invoke.ts
@@ -13,7 +13,7 @@ if (!process.env.NODE_ENV) process.env.NODE_ENV = 'development';
 
 import mri from 'mri';
 
-import type {Args} from './types.js';
+import type {Args} from './index.js';
 import {invokeTask} from '../task/invokeTask.js';
 
 /*
