+++ src/fs/modules.ts
@@ -7,7 +7,7 @@ import type {PathStats, PathData} from './pathData.js';
 import {toImportId, pathsFromId} from '../paths.js';
 import {UnreachableError} from '../utils/error.js';
 import {DEFAULT_BUILD_CONFIG_NAME} from '../config/defaultBuildConfig.js';
-import type {Obj, Result} from '../types.js';
+import type {Obj, Result} from '../index.js';
 
 /*
