+++ src/index.ts
@@ -5,4 +5,4 @@ export {TaskError} from './task/task.js';
 
 export type {Gen} from './gen/gen.js';
 
-export * from './types.js';
+export * from './utils/types.js';
