+++ src/project/rollup-plugin-output-css.ts
@@ -5,9 +5,9 @@ import sourcemapCodec from 'sourcemap-codec';
 import {blue, gray} from '../utils/terminal.js';
 import {outputFile} from '../fs/nodeFs.js';
 import {SystemLogger, Logger} from '../utils/log.js';
-import type {GroCssBuild, GroCssBundle} from './types.js';
+import type {GroCssBuild, GroCssBundle} from './index.js';
 import {omitUndefined} from '../utils/object.js';
-import type {PartialExcept} from '../types.js';
+import type {PartialExcept} from '../index.js';
 
 export interface Options {
 	getCssBundles(): Map<string, GroCssBundle>;
