+++ src/utils/object.ts
@@ -1,4 +1,4 @@
-import type {OmitStrict} from '../types.js';
+import type {OmitStrict} from '../index.js';
 
 // Iterated keys in `for..in` are always returned as strings,
 // so to prevent usage errors the key type of `mapFn` is always a string.
